+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## New functions / features
 
+- `distribution_binomial()` for perfect binomial distributions
 - `simulate_correlation()` Simulate correlated datasets
 - `p_significance()` Compute the probability of Practical Significance (ps)
 - `overlap()` Compute overlap between two empirical distributions